python-bot-telegram/requirements.txt
python-telegram-bot==21.9